Move minimal-json encoder/decoder implementation in its own package.
